Update JSSComputerAttributeType and add validation to mapping form
Add a class variale xpath_needs_key to JSSComputerAttributeType in order
to flag if a key is required for a valid xpath expression. E.g. for an
Extenstion attribute we need the name to pull out the correct attribute,
so xpath_needs_key would be true; for a site we only need the value to check
against.

Add some form validation to the mapping form: check the manifest and catalog
names, and check for valid combinations of fields based on manifest type. Also
check that key is set if it is required